requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
supabase>=2.3.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
google-genai>=0.5.0
langgraph>=0.2.0,<0.4.0
websockets>=12.0
httpx>=0.26.0
numpy>=1.26.0
